toast added

+++ src/pages/Authentication/SignUp.jsx
@@ -1,56 +1,59 @@
-import React ,{useState}from "react";
+import React, { useState } from "react";
 import { NavLink } from "react-router-dom";
 import "./Auth.css";
 import FormInputText from "../../components/FormInputText/FormInputText";
 import Navigation from "../../components/Navigation/Navigation";
 import Footer from "../../components/Footer/Footer";
-import axios from 'axios'
-import {  toast } from 'react-toastify';
-const SignUp = () => {
+import axios from "axios";
+import { toast } from "react-toastify";
 
-  const [fullName, setFullName] = useState("") 
-  const [email, setEmail] = useState("") 
-  const [password, setPassword] = useState("") 
-  const [confirmPassword, setconfirmPassword] = useState("") 
- 
+const EMAIL_REGEX = /^[a-zA-Z0-9+_.-]+@[a-zA-Z0-9.-]+$/;
+const SignUp = () => {
+  const [fullName, setFullName] = useState("");
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const [confirmPassword, setconfirmPassword] = useState("");
+
   const SignupHandler = async () => {
     try {
+      if (password.length === 0 || fullName.length === 0) {
+        toast.info("Credentials can't be blank");
+        return;
+      }
+
+      if (!EMAIL_REGEX.test(email)) {
+        toast.info("Please enter a valid email address");
+        return;
+      }
+
+      if (confirmPassword !== password) {
+        toast.info("Passwords Don't Match");
+        return;
+      }
       axios
-      .post(
-        "/api/auth/signup",
-        {
-          email,
-          password,
-          fullName,
-        },
-        {
-          headers: {
-            "Content-Type": "application/json",
+        .post(
+          "/api/auth/signup",
+          {
+            email,
+            password,
+            fullName,
           },
-        }
-      )
-      .then((response) => {
-        console.log(response)
-        localStorage.setItem("token", response.data.encodedToken);
-        toast.success('🦄 Wow so easy!', {
-          position: "bottom-left",
-          autoClose: 10,
-          hideProgressBar: false,
-          closeOnClick: true,
-          pauseOnHover: true,
-          draggable: true,
-          progress: undefined,
-          theme: "dark",
-          transition: "Bounce",
-          });
-      });
+          {
+            headers: {
+              "Content-Type": "application/json",
+            },
+          }
+        )
+        .then((response) => {
+          console.log(response);
+          localStorage.setItem("token", response.data.encodedToken);
+          toast.success(`Welcome, ${response.data.createdUser.userName}`);
+        });
     } catch (error) {
-      
+      toast.error("Something went wrong.");
     }
-    
   };
-  
-  
+
   return (
     <div>
       <Navigation />
@@ -62,8 +65,8 @@ const SignUp = () => {
             placeholder={"John"}
             type={"text"}
             value={fullName}
-            onChange={(e)=>{
-              setFullName(e.target.value.trim())
+            onChange={(e) => {
+              setFullName(e.target.value.trim());
             }}
           />
 
@@ -72,8 +75,8 @@ const SignUp = () => {
             placeholder={"Johndoe@gmail.com"}
             type={"email"}
             value={email}
-            onChange={(e)=>{
-              setEmail(e.target.value.trim())
+            onChange={(e) => {
+              setEmail(e.target.value.trim());
             }}
           />
           <FormInputText
@@ -81,8 +84,8 @@ const SignUp = () => {
             placeholder={"*******"}
             type={"password"}
             value={password}
-            onChange={(e)=>{
-              setPassword(e.target.value.trim())
+            onChange={(e) => {
+              setPassword(e.target.value.trim());
             }}
           />
           <FormInputText
@@ -90,23 +93,31 @@ const SignUp = () => {
             placeholder={"*******"}
             type={"password"}
             value={confirmPassword}
-            onChange={(e)=>{
-              setconfirmPassword(e.target.value.trim())
+            onChange={(e) => {
+              setconfirmPassword(e.target.value.trim());
             }}
           />
           <div className="check-input">
-         <label htmlFor="">
-        <input type="checkbox" /> I accept all terms and conditions
-        </label>
-        </div>
+            <label htmlFor="">
+              <input type="checkbox" /> I accept all terms and conditions
+            </label>
+          </div>
           <div className="btn">
-            <NavLink >
-              <button className="primary_btn" onClick={()=>{
-                SignupHandler()
-              }}>SignUp</button>
+            <NavLink>
+              <button
+                className="primary_btn"
+                onClick={() => {
+                  SignupHandler();
+                }}
+              >
+                SignUp
+              </button>
+            </NavLink>
+
+            <NavLink to="/login" className="create__new__btn">
+              <p>Already have an Account? ❯ </p>
             </NavLink>
           </div>
-         
         </form>
       </main>
       <Footer />
